Basic idea of a board and a script to generates one
There actually is no board yet. There is a game, which
is a collection of operations along with dimensions of
the board. An operation is an action taken in the game.
A posn is a location on the board.

There are only four actions in the game: Add/Remove Tree/Rock.

gen_game generates a list of random actions for a fixed size game