Getting the basic ideas down
There is now a Dynamo which runs the game and hopes to be the only
thing that mutates any state. There is a stateful board object which
is used by the Dynamo to record the current state of the game.
The board has tiles on it which record what is on that board area.

Unit tests to ensure the Dynamo can run through a list of add/remove
tree/rock commands and correctly update the board state.