Add real time to Game_ops
Future commits will make use of this to allow interactions that
take place over periods of time (eg buildling something that takes
60s of real time).

BREAKING this is incompatible with all previous serialized board
files.

Games assert that the ops that are added are monotonically increasing
This seems important to preserve sanity as time goes on.